+++ corbeans-corda-webserver/src/main/kotlin/com/github/manosbatsis/corbeans/corda/webserver/components/EchoController.kt
@@ -24,13 +24,11 @@ import net.corda.core.crypto.SecureHash
 import net.corda.core.identity.CordaX500Name
 import org.apache.logging.log4j.LogManager
 import org.springframework.web.bind.annotation.*
-import springfox.documentation.annotations.ApiIgnore
 
 /**
  * Used for conversion/formatter testing
  */
 @RestController
-@ApiIgnore
 @RequestMapping("api/echo")
 class EchoController {